+++ version/changelog.md
@@ -9,6 +9,8 @@ ou can find the binary downloads here: https://github.com/NHendricks/nh-toolbox/
 
 (you can download and test the dev version from https://github.com/NHendricks/nice-2-have-toolbox/actions/workflows/node.js.yml)
 
+- export / import restic settings
+
 ### v1.8.0 (10.02.2026)
 
 - build linux app